Crate kumo_prometheus

Re-exports

pub use paste;

Modules

counter_bundle

labels

parser

registry

Macros

count_tts

macro_rules! implementation of count_tts. Source: https://github.com/camsteffen/count-tts

counter_bundle

counter_bundle declares a struct holding a bundle of counters that should be incremented or decremented together. This is used to facilitate computing rolled up metrics.

label_key

Used to declare a label key struct suitable for use in registering counters in the counter registry.

Structs

AtomicCounter

CounterRegistry

Either a Counter or Gauge with a specific name, where there can be multiple labelled counter instances.

WeakAtomicCounter

Traits

AtomicCounterEntry

This trait enables having prunable and non-pruning value types in the CounterRegistry.

Type Aliases

PruningCounterRegistry