kumo_prometheus::registry

Trait StreamingCollector

pub trait StreamingCollector {
    // Required methods
    fn stream_text(&self, prefix: &Option<String>) -> BoxStream<'_, String>;
    fn stream_json(&self) -> BoxStream<'_, String>;
    fn prune(&self);
}

Required Methods

fn stream_text(&self, prefix: &Option<String>) -> BoxStream<'_, String>

Stream chunks of text in prometheus text exposition format

fn stream_json(&self) -> BoxStream<'_, String>

Stream chunks in our json format, as chunks of text

fn prune(&self)

Prune any stale entries from this collector