kumo_prometheus
Struct CounterRegistry

pub struct CounterRegistry<K, V: AtomicCounterEntry = AtomicCounter> { /* private fields */ }

Either a Counter or Gauge with a specific name, where there can be multiple labelled counter instances.

CounterRegistry has a PruningCounterRegistry variant which will drop unreferenced counter instances when they fall out of scope.

The key type K must be created via the label_key! macro provided by this crate. It allows making type-safe keys and resolving counter instances without making extraneous copies of the keys.

CounterRegistry implements the StreamingCollector trait which allows for efficient streaming serialization of its set of counters in either text or json format.

Implementations

impl<K: Clone + Send + Sync + MetricLabel + 'static, V: AtomicCounterEntry + 'static> CounterRegistry<K, V>

pub fn register(name: &'static str, help: &'static str) -> Self

Register a set of Counters, values that are only allowed to increment.

pub fn register_gauge(name: &'static str, help: &'static str) -> Self

Register a set of Gauges, values that are allowed to increase and decrease.

impl<K, V> CounterRegistry<K, V>
where V: AtomicCounterEntry, K: Eq + Hash + MetricLabel,

pub fn get<Q>(&self, key: &Q) -> Option<AtomicCounter>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Resolve an already-existing counter for the given key, or None if there either has never been such a value, or if it was pruned.

pub fn get_or_create<'a, Q>(&self, key: &'a Q) -> AtomicCounter
where K: Borrow<Q> + From<&'a Q>, Q: Hash + Eq + ?Sized,

Resolve an already-existing counter for the given key, creating a new one if it didnโt already exist, or was previously pruned.