kumo_prometheus
Trait AtomicCounterEntry

pub trait AtomicCounterEntry: Send + Sync {
    // Required methods
    fn resolve(&self) -> Option<AtomicCounter>;
    fn make_storable(strong: &AtomicCounter) -> Self;
    fn needs_pruning() -> bool;
}

This trait enables having prunable and non-pruning value types in the CounterRegistry.

Required Methods

fn resolve(&self) -> Option<AtomicCounter>

resolve this entry to an AtomicCounter instance

fn make_storable(strong: &AtomicCounter) -> Self

Given a new strong AtomicCounter reference, return Self suitable for storing in the counter registry

fn needs_pruning() -> bool

Indicate whether this type of value requires pruning the containing counter registry

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.