Crate kumo_prometheus

Re-exports

pub use pastey as paste;

pub use prometheus;

Modules

counter_bundle

labels

parser

registry

Macros

count_tts

macro_rules! implementation of count_tts. Source: https://github.com/camsteffen/count-tts

counter_bundle

counter_bundle declares a struct holding a bundle of counters that should be incremented or decremented together. This is used to facilitate computing rolled up metrics.

declare_metric

This macro aids in declaring metrics. Usage looks like:

label_key

Used to declare a label key struct suitable for use in registering counters in the counter registry.

mandatory_doc

This macro matches a series of doc comment attributes. Multi-line doc comments appear as a sequence of doc comment attributes, so we need to be able to match both the individual case and the sequence, and map them back to a single string.

Structs

AtomicCounter

CounterDescription

CounterRegistry

Either a Counter or Gauge with a specific name, where there can be multiple labelled counter instances.

WeakAtomicCounter

Enums

MetricPrune

MetricType

Statics

COUNTER_METADATA

Accumulates metric metadata

Traits

AtomicCounterEntry

This trait enables having prunable and non-pruning value types in the CounterRegistry.

Functions

export_metadata

split_help

Utility function for dealing with doc comment metadata. Look for two successive line breaks; if they are present they denote the break between the short first-logical-line and a longer descriptive exposition. Returns that first logical line and the optional exposition.

Type Aliases

PruningCounterRegistry